scripts/requirements.txt
# Vahan Analytics - Data Fetch Scripts
# pip install -r requirements.txt

requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
tabulate>=0.9.0
